chore: rename util package to utils and update all imports
